tweak robo-browser.py to add stability and prevent multiple instances from fighting over the same robotask
from this point on, the robo-browser is invoked from the bash script, which will help prevent runaway
chrome tasks from leaking memory. when getting the next robotask, we now pick randomly from the next 10 tasks.
this is a temporary fix for the problem where multiple robo-browsers would grab a single task and then
fail on teh DELETE call